app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
